+++ HISTORY.rst
@@ -7,6 +7,7 @@ History
 * Add support for 15 min data resolution
 * Change group_by_dynamic to include 'closed='left'' and 'label' parameters for better resampling control
 * Add resample_data_by_time_step utility function
+* Set default min_count to 50% of expected values when resampling data
 * Update neighbourhood checks to handle 15 min data and resampling
 
 0.3.2 (2025-01-22)